Move ContentConsistencyChecker into DataHygiene module
