+++ lib/tasks/check_content_consistency.rake
@@ -1,8 +1,8 @@
-require "content_consistency_checker"
+require "data_hygiene/content_consistency_checker"
 
 def report_errors(errors, content_store)
   Airbrake.notify(
-    "Inconsistent #{content_store} documents",
+    "Documents inconsistent with the #{content_store} content store",
     parameters: {
       errors: errors,
     }
@@ -14,7 +14,7 @@ def report_errors(errors, content_store)
   end
 end
 
-desc "Check all the documents for consistency with the router-api and content-store"
+desc "Check all the documents for consistency with the content-store"
 task :check_content_consistency, [:content_store, :content_dump] => [:environment] do |_, args|
   raise "Missing content store." unless args[:content_store]
   raise "Invalid content store." unless %w(live draft).include?(args[:content_store])
@@ -22,7 +22,7 @@ task :check_content_consistency, [:content_store, :content_dump] => [:environmen
 
   content_dump = ContentDumpLoader.load(args[:content_dump])
 
-  checker = ContentConsistencyChecker.new(args[:content_store], content_dump)
+  checker = DataHygiene::ContentConsistencyChecker.new(args[:content_store], content_dump)
   checker.check_editions
   checker.check_content
